+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Log request parameters for Rodauth requests (@janko)
+
 ## 0.11.0 (2021-05-06)
 
 * Add controller-like logging for requests to Rodauth endpoints (@janko)